home and account done, working on profile

+++ src/main/java/com/transaction/Transaction.java
@@ -151,7 +151,7 @@ public class Transaction {
         this.paymentMethod = paymentMethod;
     }
 
-    public float getBalanceBefore() {
+    public double getBalanceBefore() {
         return user.getBalance();
     }
